Move `InfoGraphDot` into its own module.
